requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.61.1
django-cors-headers==4.4.0
django-filter==24.2
django-grappelli==4.0.1
django-import-export==4.0.9
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
importlib_metadata==7.2.0
Markdown==3.6
mysqlclient==2.2.4
ndg-httpsclient==0.5.1
oauthlib==3.2.2
pillow==10.3.0
pusher==3.3.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
zipp==3.19.2
